+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased 
 * Namicorn#ipfsHash -> gets ipfs hash for a specific supported domain
-* Namicorn#ipfsEmail -> gets ipfs email field of whois object for a specific supported domain
+* Namicorn#email -> gets ipfs email field of whois object for a specific supported domain
 * Namicorn#ipfsRedirect -> gets ipfs redirect url record for a specific supported domain
 
 ## 0.3.1 - 0.3.2